Combine cenv and system to cstdlib
